Unity 2022.2.0a10 C# reference source code

+++ Modules/QuickSearch/Editor/SearchService.cs
@@ -575,7 +575,8 @@ namespace UnityEditor.Search
         /// <returns></returns>
         public static ISearchView ShowWindow(SearchViewState viewState)
         {
-            return QuickSearch.Create(viewState).ShowWindow();
+            return QuickSearch.Create(viewState).ShowWindow(viewState.windowSize.x, viewState.windowSize.y,
+                SearchFlags.OpenDefault | (viewState.context?.options ?? SearchFlags.None));
         }
 
         /// <summary>
@@ -647,7 +648,7 @@ namespace UnityEditor.Search
             return SearchPickerWindow.ShowPicker(viewState);
         }
 
-        internal static IEnumerable<SearchProvider> GetActiveProviders()
+        public static IEnumerable<SearchProvider> GetActiveProviders()
         {
             return Providers.Where(p => p.active);
         }
@@ -727,6 +728,12 @@ namespace UnityEditor.Search
             TrackCreateIndex(db, options, indexName, indexPath, onIndexReady, 1d);
         }
 
+        public static IEnumerable<ISearchDatabase> EnumerateDatabases()
+        {
+            foreach (var db in SearchDatabase.EnumerateAll())
+                yield return db;
+        }
+
         /// <summary>
         /// Checks if a search index is ready to be used.
         /// </summary>
@@ -752,10 +759,23 @@ namespace UnityEditor.Search
             {
                 onIndexReady?.Invoke(indexName, indexPath.Replace("\\", "/"), () =>
                 {
-                    if (EditorUtility.IsPersistent(db))
-                        Resources.UnloadAsset(db);
-                    if (options.HasNone(IndexingOptions.Keep))
-                        AssetDatabase.DeleteAsset(indexPath);
+                    SearchDatabase.Unload(db);
+
+                    try
+                    {
+                        if (options.HasNone(IndexingOptions.Keep) && System.IO.File.Exists(indexPath))
+                        {
+                            if (options.HasAny(IndexingOptions.Temporary))
+                                System.IO.File.Delete(indexPath);
+                            else
+                                AssetDatabase.DeleteAsset(indexPath);
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        // Ignore any file IO errors (for the user)
+                        Console.WriteLine(ex.Message);
+                    }
                 });
             }
             else
@@ -824,7 +844,7 @@ namespace UnityEditor.Search
             }
             catch (SearchExpressionEvaluatorException ex)
             {
-                var queryError = new SearchQueryError(ex.errorView.startIndex, ex.errorView.Length, ex.Message,
+                var queryError = new SearchQueryError(ex.errorView.startIndex, ex.errorView.length, ex.Message,
                     context, s_SearchServiceProvider, fromSearchQuery: false, SearchQueryErrorType.Error);
                 context.AddSearchQueryError(queryError);
                 return false;
